refactor(youtube/litho-filter): move classes to their respective file
